ensure all jupyter kernels are closed on project exit; refactor some jupyter exit code
- there are many settings, e.g., kucalc, where cleaning up all kernels
  on project stop (or restart) isn't important.  BUT there are settings
  where it *is* critically important, e.g., when doing development and
  possibly on cocalc-docker in some cases now that users aren't removed.
  This improvement does a good job of ensuring that all the kernels that
  a project spawns are cleaned up when it terminates.
